add in initial support for fixed length things unclear if the interface is now good will move the fixed annotation to the class level, can validate things there
