+++ atlasdb-schema/src/main/java/com/palantir/atlasdb/schema/apt/KeyDefinition.java
@@ -21,4 +21,7 @@ import org.immutables.value.Value;
 public interface KeyDefinition {
 	String getName();
 	String getKeyTypeFullyQualified();
+	
+	boolean isFixed();
+	int getLength();
 }
